Modified helper package to include send empty raw error (#1495)

+++ gateway/modules/schema/inspection_test.go
@@ -79,7 +79,7 @@ func Test_generateInspection(t *testing.T) {
 			args: args{
 				dbType: "mysql",
 				col:    "table1",
-				fields: []model.InspectorFieldType{{ColumnName: firstColumn, FieldType: "float", FieldNull: "YES", NumericPrecision: 10, NumericScale: 5}},
+				fields: []model.InspectorFieldType{{ColumnName: firstColumn, FieldType: "decimal", FieldNull: "YES", NumericPrecision: 10, NumericScale: 5}},
 			},
 			want:    model.Collection{"table1": model.Fields{firstColumn: &model.FieldType{FieldName: firstColumn, Kind: model.TypeFloat, Args: &model.FieldArgs{Precision: 10, Scale: 5}}}},
 			wantErr: false,
@@ -158,7 +158,7 @@ func Test_generateInspection(t *testing.T) {
 			args: args{
 				dbType: "mysql",
 				col:    "table1",
-				fields: []model.InspectorFieldType{{ColumnName: firstColumn, FieldType: "float", FieldNull: "NO", NumericPrecision: 10, NumericScale: 5}},
+				fields: []model.InspectorFieldType{{ColumnName: firstColumn, FieldType: "decimal", FieldNull: "NO", NumericPrecision: 10, NumericScale: 5}},
 			},
 			want:    model.Collection{"table1": model.Fields{firstColumn: &model.FieldType{FieldName: firstColumn, IsFieldTypeRequired: true, Kind: model.TypeFloat, Args: &model.FieldArgs{Precision: 10, Scale: 5}}}},
 			wantErr: false,
@@ -252,7 +252,7 @@ func Test_generateInspection(t *testing.T) {
 			args: args{
 				dbType: "mysql",
 				col:    "table1",
-				fields: []model.InspectorFieldType{{ColumnName: firstColumn, FieldType: "float", FieldNull: "NO", FieldDefault: "9.8", AutoIncrement: "false", NumericPrecision: 10, NumericScale: 5}},
+				fields: []model.InspectorFieldType{{ColumnName: firstColumn, FieldType: "decimal", FieldNull: "NO", FieldDefault: "9.8", AutoIncrement: "false", NumericPrecision: 10, NumericScale: 5}},
 			},
 			want:    model.Collection{"table1": model.Fields{firstColumn: &model.FieldType{FieldName: firstColumn, IsFieldTypeRequired: true, IsDefault: true, Kind: model.TypeFloat, Args: &model.FieldArgs{Precision: 10, Scale: 5}, Default: "9.8"}}},
 			wantErr: false,
@@ -385,7 +385,7 @@ func Test_generateInspection(t *testing.T) {
 			args: args{
 				dbType:    "mysql",
 				col:       "table1",
-				fields:    []model.InspectorFieldType{{ColumnName: firstColumn, FieldType: "float", FieldNull: "NO", NumericPrecision: 10, NumericScale: 5}},
+				fields:    []model.InspectorFieldType{{ColumnName: firstColumn, FieldType: "decimal", FieldNull: "NO", NumericPrecision: 10, NumericScale: 5}},
 				indexKeys: []model.IndexType{{TableName: "table1", ColumnName: firstColumn, IndexName: getIndexName("table1", "index1"), Order: 1, Sort: model.DefaultIndexSort, IsUnique: true}},
 			},
 			want:    model.Collection{"table1": model.Fields{firstColumn: &model.FieldType{FieldName: firstColumn, IsFieldTypeRequired: true, Args: &model.FieldArgs{Precision: 10, Scale: 5}, Kind: model.TypeFloat, IndexInfo: []*model.TableProperties{{Field: firstColumn, IsUnique: true, Group: "index1", Order: 1, Sort: model.DefaultIndexSort, ConstraintName: getIndexName("table1", "index1")}}}}},
@@ -480,7 +480,7 @@ func Test_generateInspection(t *testing.T) {
 			args: args{
 				dbType: "mysql",
 				col:    "table1",
-				fields: []model.InspectorFieldType{{ColumnName: firstColumn, FieldType: "float", FieldNull: "NO", NumericPrecision: 10, NumericScale: 5}, {ColumnName: secondColumn, FieldType: "float", FieldNull: "NO", NumericPrecision: 10, NumericScale: 5}},
+				fields: []model.InspectorFieldType{{ColumnName: firstColumn, FieldType: "decimal", FieldNull: "NO", NumericPrecision: 10, NumericScale: 5}, {ColumnName: secondColumn, FieldType: "decimal", FieldNull: "NO", NumericPrecision: 10, NumericScale: 5}},
 				indexKeys: []model.IndexType{
 					{TableName: "table1", ColumnName: firstColumn, IndexName: getIndexName("table1", "index1"), Order: 1, Sort: model.DefaultIndexSort, IsUnique: true},
 					{TableName: "table1", ColumnName: secondColumn, IndexName: getIndexName("table1", "index1"), Order: 2, Sort: model.DefaultIndexSort, IsUnique: true},
@@ -581,7 +581,7 @@ func Test_generateInspection(t *testing.T) {
 			args: args{
 				dbType:    "mysql",
 				col:       "table1",
-				fields:    []model.InspectorFieldType{{ColumnName: firstColumn, FieldType: "float", FieldNull: "NO", NumericPrecision: 10, NumericScale: 5}},
+				fields:    []model.InspectorFieldType{{ColumnName: firstColumn, FieldType: "decimal", FieldNull: "NO", NumericPrecision: 10, NumericScale: 5}},
 				indexKeys: []model.IndexType{{TableName: "table1", ColumnName: firstColumn, IndexName: "custom-index", Order: 1, Sort: model.DefaultIndexSort, IsUnique: true}},
 			},
 			want:    model.Collection{"table1": model.Fields{firstColumn: &model.FieldType{FieldName: firstColumn, IsFieldTypeRequired: true, Kind: model.TypeFloat, Args: &model.FieldArgs{Precision: 10, Scale: 5}, IndexInfo: []*model.TableProperties{{Field: firstColumn, IsUnique: true, Group: "custom-index", Order: 1, Sort: model.DefaultIndexSort, ConstraintName: "custom-index"}}}}},
@@ -676,7 +676,7 @@ func Test_generateInspection(t *testing.T) {
 			args: args{
 				dbType: "mysql",
 				col:    "table1",
-				fields: []model.InspectorFieldType{{ColumnName: firstColumn, FieldType: "float", FieldNull: "NO", NumericPrecision: 10, NumericScale: 5}, {ColumnName: secondColumn, FieldType: "float", FieldNull: "NO", NumericPrecision: 10, NumericScale: 5}},
+				fields: []model.InspectorFieldType{{ColumnName: firstColumn, FieldType: "decimal", FieldNull: "NO", NumericPrecision: 10, NumericScale: 5}, {ColumnName: secondColumn, FieldType: "decimal", FieldNull: "NO", NumericPrecision: 10, NumericScale: 5}},
 				indexKeys: []model.IndexType{
 					{TableName: "table1", ColumnName: firstColumn, IndexName: "custom-index", Order: 1, Sort: model.DefaultIndexSort, IsUnique: true},
 					{TableName: "table1", ColumnName: secondColumn, IndexName: "custom-index", Order: 2, Sort: model.DefaultIndexSort, IsUnique: true},
@@ -777,7 +777,7 @@ func Test_generateInspection(t *testing.T) {
 			args: args{
 				dbType:    "mysql",
 				col:       "table1",
-				fields:    []model.InspectorFieldType{{ColumnName: firstColumn, FieldType: "float", FieldNull: "NO", NumericPrecision: 10, NumericScale: 5}},
+				fields:    []model.InspectorFieldType{{ColumnName: firstColumn, FieldType: "decimal", FieldNull: "NO", NumericPrecision: 10, NumericScale: 5}},
 				indexKeys: []model.IndexType{{TableName: "table1", ColumnName: firstColumn, IndexName: getIndexName("table1", "index1"), Order: 1, Sort: model.DefaultIndexSort, IsUnique: false}},
 			},
 			want:    model.Collection{"table1": model.Fields{firstColumn: &model.FieldType{FieldName: firstColumn, IsFieldTypeRequired: true, Args: &model.FieldArgs{Precision: 10, Scale: 5}, Kind: model.TypeFloat, IndexInfo: []*model.TableProperties{{Field: firstColumn, IsIndex: true, Group: "index1", Order: 1, Sort: model.DefaultIndexSort, ConstraintName: getIndexName("table1", "index1")}}}}},
@@ -872,7 +872,7 @@ func Test_generateInspection(t *testing.T) {
 			args: args{
 				dbType: "mysql",
 				col:    "table1",
-				fields: []model.InspectorFieldType{{ColumnName: firstColumn, FieldType: "float", FieldNull: "NO", NumericPrecision: 10, NumericScale: 5}, {ColumnName: secondColumn, FieldType: "float", FieldNull: "NO", NumericPrecision: 10, NumericScale: 5}},
+				fields: []model.InspectorFieldType{{ColumnName: firstColumn, FieldType: "decimal", FieldNull: "NO", NumericPrecision: 10, NumericScale: 5}, {ColumnName: secondColumn, FieldType: "decimal", FieldNull: "NO", NumericPrecision: 10, NumericScale: 5}},
 				indexKeys: []model.IndexType{
 					{TableName: "table1", ColumnName: firstColumn, IndexName: getIndexName("table1", "index1"), Order: 1, Sort: model.DefaultIndexSort, IsUnique: false},
 					{TableName: "table1", ColumnName: secondColumn, IndexName: getIndexName("table1", "index1"), Order: 2, Sort: model.DefaultIndexSort, IsUnique: false},
@@ -973,7 +973,7 @@ func Test_generateInspection(t *testing.T) {
 			args: args{
 				dbType:    "mysql",
 				col:       "table1",
-				fields:    []model.InspectorFieldType{{ColumnName: firstColumn, FieldType: "float", FieldNull: "NO", NumericPrecision: 10, NumericScale: 5}},
+				fields:    []model.InspectorFieldType{{ColumnName: firstColumn, FieldType: "decimal", FieldNull: "NO", NumericPrecision: 10, NumericScale: 5}},
 				indexKeys: []model.IndexType{{TableName: "table1", ColumnName: firstColumn, IndexName: "custom-index", Order: 1, Sort: model.DefaultIndexSort, IsUnique: false}},
 			},
 			want:    model.Collection{"table1": model.Fields{firstColumn: &model.FieldType{FieldName: firstColumn, IsFieldTypeRequired: true, Kind: model.TypeFloat, Args: &model.FieldArgs{Precision: 10, Scale: 5}, IndexInfo: []*model.TableProperties{{Field: firstColumn, IsIndex: true, Group: "custom-index", Order: 1, Sort: model.DefaultIndexSort, ConstraintName: "custom-index"}}}}},
@@ -1068,7 +1068,7 @@ func Test_generateInspection(t *testing.T) {
 			args: args{
 				dbType: "mysql",
 				col:    "table1",
-				fields: []model.InspectorFieldType{{ColumnName: firstColumn, FieldType: "float", FieldNull: "NO", NumericPrecision: 10, NumericScale: 5}, {ColumnName: secondColumn, FieldType: "float", FieldNull: "NO", NumericPrecision: 10, NumericScale: 5}},
+				fields: []model.InspectorFieldType{{ColumnName: firstColumn, FieldType: "decimal", FieldNull: "NO", NumericPrecision: 10, NumericScale: 5}, {ColumnName: secondColumn, FieldType: "decimal", FieldNull: "NO", NumericPrecision: 10, NumericScale: 5}},
 				indexKeys: []model.IndexType{
 					{TableName: "table1", ColumnName: firstColumn, IndexName: "custom-index", Order: 1, Sort: model.DefaultIndexSort, IsUnique: false},
 					{TableName: "table1", ColumnName: secondColumn, IndexName: "custom-index", Order: 2, Sort: model.DefaultIndexSort, IsUnique: false},
@@ -1178,7 +1178,7 @@ func Test_generateInspection(t *testing.T) {
 			args: args{
 				dbType: "postgres",
 				col:    "table1",
-				fields: []model.InspectorFieldType{{ColumnName: firstColumn, FieldType: "float", FieldNull: "NO", FieldDefault: "9.8", AutoIncrement: "false"}},
+				fields: []model.InspectorFieldType{{ColumnName: firstColumn, FieldType: "decimal", FieldNull: "NO", FieldDefault: "9.8", AutoIncrement: "false"}},
 			},
 			want:    model.Collection{"table1": model.Fields{firstColumn: &model.FieldType{FieldName: firstColumn, IsFieldTypeRequired: true, Kind: model.TypeFloat, IsDefault: true, Default: "9.8"}}},
 			wantErr: false,
@@ -1278,7 +1278,7 @@ func Test_generateInspection(t *testing.T) {
 			args: args{
 				dbType: "sqlserver",
 				col:    "table1",
-				fields: []model.InspectorFieldType{{ColumnName: firstColumn, FieldType: "float", FieldNull: "NO", FieldDefault: "9.8", AutoIncrement: "false"}},
+				fields: []model.InspectorFieldType{{ColumnName: firstColumn, FieldType: "decimal", FieldNull: "NO", FieldDefault: "9.8", AutoIncrement: "false"}},
 			},
 			want:    model.Collection{"table1": model.Fields{firstColumn: &model.FieldType{FieldName: firstColumn, IsFieldTypeRequired: true, IsDefault: true, Kind: model.TypeFloat, Default: "9.8"}}},
 			wantErr: false,
